chore: moved fileType to helper
