+++ src/helper/index.ts
@@ -119,3 +119,23 @@ export async function commandInput(obj: {
 
   return resultString;
 }
+
+export async function fileType(cmd: string): Promise<string> {
+  const formatInput = await vscode.window.showQuickPick(
+    [{ label: 'json', picked: true }, { label: 'yaml/yml' }, { label: 'url' }, { label: 'base64' }],
+    { title: `${cmd} format`, placeHolder: 'Set the output format', ignoreFocusOut: true }
+  );
+  if (formatInput === undefined) {
+    return 'noUploadTypeSelected';
+  }
+  switch (formatInput?.label) {
+    case 'yaml/yml':
+      return 'yaml';
+    case 'url':
+      return 'url';
+    case 'base64':
+      return 'base64';
+    default:
+      return 'json';
+  }
+}
