+++ src/oryUpdate.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { spawn } from 'child_process';
 import { outputChannel, oryCommand } from './extension';
-import { spawnCommonErrAndClose } from './helper';
+import { spawnCommonErrAndClose, fileType } from './helper';
 import { oauth2Client } from './oryCreate';
 
 export async function runOryUpdate() {
@@ -247,26 +247,6 @@ async function getInputBox(title: string, placeHolder: string) {
   });
 
   return userInput;
-}
-
-export async function fileType(cmd: string): Promise<string> {
-  const formatInput = await vscode.window.showQuickPick(
-    [{ label: 'json', picked: true }, { label: 'yaml/yml' }, { label: 'url' }, { label: 'base64' }],
-    { title: `${cmd} format`, placeHolder: 'Set the output format', ignoreFocusOut: true }
-  );
-  if (formatInput === undefined) {
-    return 'noUploadTypeSelected';
-  }
-  switch (formatInput?.label) {
-    case 'yaml/yml':
-      return 'yaml';
-    case 'url':
-      return 'url';
-    case 'base64':
-      return 'base64';
-    default:
-      return 'json';
-  }
 }
 
 export async function commandInput(obj: {
